Bug 572532: [Tests] UML validateCamelCaseRule test fails consistently
OCL Pivot maintains environments on a per-thread basis, which means the
Modal Context thread cannot use an environment set up by pre-validation
hooks on the main thread. Thus this pre-processing needs to be performed
on the Modal Context thread also and the test needs to be updated to
use the hook mechanism to inject its OCL constraints from the file.

Change-Id: Id3fdb3b9f8ab876d9ad6a7b39500684c18133f28
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>