Bug 573656: [Welcome Page] Concurrent Modification issue on Welcome page while generating Element Types for a profile
Privileged runnables cannot be run asynchronously, otherwise there
is every possibility that the transaction that owns them will no
longer be active when the runnable runs. So, instead, create a new
read-only transaction in the asynchronous runnable to process the
notifications and update the observable property on the UI thread.

Change-Id: I35ce6845415d8a90108bca5a1877c7c0b11b747a
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>