Bug 574094: [Toolsmiths - Properties Editor] Unmodified XWT files are saved/reformatted
The mechanism of trying to create an output stream to test for writability of a resource
truncates the resource to 0 bytes, after which it is (usually) saved again with full
content. But that defeats EMF's save-only-if-modified-by-memory-buffer strategy.

Instead, give the editing domain specific knowledge of ppe: URI scheme and use
heuristics supported by EMF (including URI schemes and URIConverter) to test for
real read-only condition and also, while saving, provide ephemeral read-only state
for resources that are not modified. This prevents unmodified but otherwise writable
XWT resources being saved again with slightly rearranged but otherwise identical content.

Change-Id: I0df7e37847df85eb5a0f84f9b1feae7020170be8
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>